+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/combat/fletching/Fletching.java
@@ -17,7 +17,6 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
-import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.world.InteractionHand;
@@ -27,11 +26,9 @@ import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockBehaviour;
-import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.RegistryObject;
@@ -119,14 +116,5 @@ public class Fletching extends Feature {
 
 		event.getEntity().sendSystemMessage(Component.translatable(INVALID_FLETCHING_LANG));
 		event.setCanceled(true);
-	}
-
-	@SubscribeEvent
-	public void onTooltip(ItemTooltipEvent event) {
-		if (!this.isEnabled()
-				|| !event.getItemStack().is(Items.ARROW))
-			return;
-
-		event.getToolTip().add(Component.translatable(Items.ARROW.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.GRAY));
 	}
 }